APIs/configcat.com/v1/python-aiohttp-client/tests/test_members_controller.py
# coding: utf-8

import pytest
import json
from aiohttp import web

from openapi_server.models.add_user_to_group_request import AddUserToGroupRequest
from openapi_server.models.invite_members_request import InviteMembersRequest
from openapi_server.models.member_model import MemberModel
from openapi_server.models.user_model import UserModel


pytestmark = pytest.mark.asyncio

@pytest.mark.skip("Connexion does not support multiple consumes. See https://github.com/zalando/connexion/pull/760")
async def test_add_member_to_group(client):
    """Test case for add_member_to_group

    Update Member Permissions
    """
    body = {"permissionGroupIds":[0,0]}
    headers = { 
        'Content-Type': 'application/*+json',
        'Authorization': 'BasicZm9vOmJhcg==',
    }
    response = await client.request(
        method='POST',
        path='/v1/organizations/{organization_id}/members/{user_id}'.format(organization_id='organization_id_example', user_id='user_id_example'),
        headers=headers,
        json=body,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_delete_organization_member(client):
    """Test case for delete_organization_member

    Delete Member from Organization
    """
    headers = { 
        'Authorization': 'BasicZm9vOmJhcg==',
    }
    response = await client.request(
        method='DELETE',
        path='/v1/organizations/{organization_id}/members/{user_id}'.format(organization_id='organization_id_example', user_id='user_id_example'),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_delete_product_member(client):
    """Test case for delete_product_member

    Delete Member from Product
    """
    headers = { 
        'Authorization': 'BasicZm9vOmJhcg==',
    }
    response = await client.request(
        method='DELETE',
        path='/v1/products/{product_id}/members/{user_id}'.format(product_id='product_id_example', user_id='user_id_example'),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_get_organization_members(client):
    """Test case for get_organization_members

    List Organization Members
    """
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'BasicZm9vOmJhcg==',
    }
    response = await client.request(
        method='GET',
        path='/v1/organizations/{organization_id}/members'.format(organization_id='organization_id_example'),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_get_product_members(client):
    """Test case for get_product_members

    List Product Members
    """
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'BasicZm9vOmJhcg==',
    }
    response = await client.request(
        method='GET',
        path='/v1/products/{product_id}/members'.format(product_id='product_id_example'),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

@pytest.mark.skip("Connexion does not support multiple consumes. See https://github.com/zalando/connexion/pull/760")
async def test_invite_member(client):
    """Test case for invite_member

    Invite Member
    """
    body = {"emails":["emails","emails"],"permissionGroupId":0}
    headers = { 
        'Content-Type': 'application/*+json',
        'Authorization': 'BasicZm9vOmJhcg==',
    }
    response = await client.request(
        method='POST',
        path='/v1/products/{product_id}/members/invite'.format(product_id='product_id_example'),
        headers=headers,
        json=body,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')

